requirements.txt
aiounittest==1.4.2
alabaster==0.7.16
anyio==4.8.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.4
attrs==25.1.0
babel==2.16.0
beautifulsoup4==4.12.3
bleach==6.2.0
bokeh==3.6.2
capytaine==2.2.1
certifi==2024.12.14
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
coverage==7.6.10
cycler==0.12.1
debugpy==1.8.12
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.9
docopt==0.6.2
docutils==0.18.1
exceptiongroup==1.2.2
executing==2.2.0
fastjsonschema==2.21.1
filelock==3.17.0
fonttools==4.55.3
fqdn==1.5.1
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
identify==2.6.6
idna==3.10
imagesize==1.4.1
importlib_metadata==8.6.1
ipykernel==6.29.5
ipympl==0.9.6
ipyparallel==9.0.0
ipython==8.31.0
ipywidgets==8.1.5
isoduration==20.11.0
jax==0.5.0
jaxlib==0.5.0
jedi==0.19.2
Jinja2==3.1.5
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-book==0.14.0
jupyter-cache==0.6.1
jupyter-events==0.11.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kiwisolver==1.4.8
latexcodec==3.0.0
linkify-it-py==2.0.3
Markdown==3.7
markdown-it-py==2.2.0
MarkupSafe==3.0.2
matlabengine==24.2.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdit-py-plugins==0.3.5
mdurl==0.1.2
mistune==3.1.1
ml_dtypes==0.5.1
myst-nb==0.17.2
myst-parser==0.18.1
nbclient==0.7.4
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
nodeenv==1.9.1
notebook==7.3.2
notebook_shim==0.2.4
num2words==0.5.14
numpy==2.2.1
numpydoc==1.6.0
omxdsm @ https://github.com/onodip/OpenMDAO-XDSM/tarball/master#sha256=3d2bd951bc106fbd88a14d3b78578ba55208c1fc59af5f32e1fcccd6575fcc4e
openmdao==3.36.0
opt_einsum==3.4.0
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
panel==1.6.0
param==2.2.0
parameterized==0.9.0
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
playwright==1.49.1
pre_commit==4.1.0
prometheus_client==0.21.1
prompt_toolkit==3.0.50
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
pybtex==0.24.0
pybtex-docutils==1.0.3
pycodestyle==2.12.1
pycparser==2.22
pydata-sphinx-theme==0.12.0
pydocstyle==6.3.0
pyDOE3==1.0.4
pyee==12.0.0
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-json-logger==3.2.1
pytz==2024.2
pyviz_comms==3.0.4
pyXDSM==2.3.1
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.36.2
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.22.3
scipy==1.15.1
Send2Trash==1.8.3
six==1.17.0
sniffio==1.3.1
snowballstemmer==2.2.0
soupsieve==2.6
Sphinx==5.0.2
sphinx-book-theme==0.4.0rc1
sphinx-comments==0.0.3
sphinx-copybutton==0.5.2
sphinx-jupyterbook-latex==0.5.2
sphinx-multitoc-numbering==0.1.3
sphinx-sitemap==2.6.0
sphinx-thebe==0.2.1
sphinx-togglebutton==0.3.2
sphinx_design==0.3.0
sphinx_external_toc==0.3.1
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-bibtex==2.5.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
SQLAlchemy==2.0.37
stack-data==0.6.3
tabulate==0.9.0
terminado==0.18.1
testflo==1.4.19
tinycss2==1.4.0
tomli==2.2.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2024.2
uc-micro-py==1.0.3
uri-template==1.3.0
urllib3==2.3.0
virtualenv==20.29.1
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==14.2
widgetsnbextension==4.0.13
wrapt==1.17.2
xarray==2025.1.1
xdsmjs==2.0.0
xyzservices==2025.1.0
zipp==3.21.0
